finished task 112

+++ src/main/java/com/example/demo/repository/ResumeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ResumeRepository extends JpaRepository<Resume,Integer> {
+public interface ResumeRepository extends JpaRepository<Resume, Long> {
 
     Page<Resume> findAll(Pageable pageable);
     List<Resume> findAllBySpecialist_Id(Long id);
